# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ivory Health Club API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MembershipPlanTier = typeof MembershipPlanTier[keyof typeof MembershipPlanTier];


export const MembershipPlanTier = {
  silver_single: 'silver_single',
  silver_family: 'silver_family',
  gold_single: 'gold_single',
  gold_family: 'gold_family',
  gold_plus: 'gold_plus',
  diamond: 'diamond',
  seventy_plus: 'seventy_plus',
} as const;

export type MembershipPlanPricePeriod = typeof MembershipPlanPricePeriod[keyof typeof MembershipPlanPricePeriod];


export const MembershipPlanPricePeriod = {
  monthly: 'monthly',
  annual: 'annual',
} as const;

export interface MembershipPlan {
  id: number;
  name: string;
  tier: MembershipPlanTier;
  description: string;
  price: number;
  pricePeriod?: MembershipPlanPricePeriod;
  perks: string[];
  discounts?: string;
  /** @nullable */
  maxMembers?: number | null;
}

export type EnrollmentStatus = typeof EnrollmentStatus[keyof typeof EnrollmentStatus];


export const EnrollmentStatus = {
  pending: 'pending',
  active: 'active',
  expired: 'expired',
  cancelled: 'cancelled',
} as const;

export type EnrollmentPaymentStatus = typeof EnrollmentPaymentStatus[keyof typeof EnrollmentPaymentStatus];


export const EnrollmentPaymentStatus = {
  unpaid: 'unpaid',
  paid: 'paid',
  failed: 'failed',
} as const;

export interface Enrollment {
  id: number;
  planId: number;
  /** @nullable */
  planName?: string | null;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  status: EnrollmentStatus;
  paymentStatus: EnrollmentPaymentStatus;
  /** @nullable */
  paymentReference?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface EnrollmentInput {
  planId: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address?: string;
  dateOfBirth?: string;
  notes?: string;
}

export interface PaymentConfirmation {
  paymentReference: string;
  notes?: string;
}

export type BookingServiceType = typeof BookingServiceType[keyof typeof BookingServiceType];


export const BookingServiceType = {
  gym: 'gym',
  spa: 'spa',
  entertainment: 'entertainment',
  restaurant: 'restaurant',
  juicebar: 'juicebar',
  event_hall: 'event_hall',
  fitness_program: 'fitness_program',
  youth_program: 'youth_program',
} as const;

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  serviceType: BookingServiceType;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  bookingDate: string;
  /** @nullable */
  bookingTime?: string | null;
  /** @nullable */
  numberOfGuests?: number | null;
  /** @nullable */
  specialRequests?: string | null;
  status: BookingStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type BookingInputServiceType = typeof BookingInputServiceType[keyof typeof BookingInputServiceType];


export const BookingInputServiceType = {
  gym: 'gym',
  spa: 'spa',
  entertainment: 'entertainment',
  restaurant: 'restaurant',
  juicebar: 'juicebar',
  event_hall: 'event_hall',
  fitness_program: 'fitness_program',
  youth_program: 'youth_program',
} as const;

export interface BookingInput {
  serviceType: BookingInputServiceType;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  bookingDate: string;
  bookingTime?: string;
  numberOfGuests?: number;
  specialRequests?: string;
}

export type BookingStatusUpdateStatus = typeof BookingStatusUpdateStatus[keyof typeof BookingStatusUpdateStatus];


export const BookingStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
  notes?: string;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  author?: string;
  published: boolean;
  createdAt: string;
}

export interface BlogPostInput {
  title: string;
  excerpt?: string;
  content: string;
  category: string;
  imageUrl?: string;
  author: string;
  published?: boolean;
}

export interface BlogPostPatch {
  title?: string;
  excerpt?: string;
  content?: string;
  category?: string;
  imageUrl?: string;
  author?: string;
  published?: boolean;
}

export interface GalleryImage {
  id: number;
  url: string;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  createdAt: string;
}

export interface GalleryImageInput {
  url: string;
  title: string;
  description?: string;
  category: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface ContactInput {
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
}

export interface MarkReadInput {
  isRead: boolean;
}

export interface AdminStats {
  totalMembers: number;
  activeMembers: number;
  pendingEnrollments: number;
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  unreadMessages: number;
  totalRevenue: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  enrollment: 'enrollment',
  booking: 'booking',
  contact: 'contact',
  payment: 'payment',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  description: string;
  timestamp: string;
  /** @nullable */
  status?: string | null;
}

export interface MembershipBreakdownItem {
  planName: string;
  tier: string;
  count: number;
}

export type ListEnrollmentsParams = {
status?: ListEnrollmentsStatus;
planId?: number;
};

export type ListEnrollmentsStatus = typeof ListEnrollmentsStatus[keyof typeof ListEnrollmentsStatus];


export const ListEnrollmentsStatus = {
  pending: 'pending',
  active: 'active',
  expired: 'expired',
  cancelled: 'cancelled',
} as const;

export type ListBookingsParams = {
status?: ListBookingsStatus;
serviceType?: ListBookingsServiceType;
};

export type ListBookingsStatus = typeof ListBookingsStatus[keyof typeof ListBookingsStatus];


export const ListBookingsStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export type ListBookingsServiceType = typeof ListBookingsServiceType[keyof typeof ListBookingsServiceType];


export const ListBookingsServiceType = {
  gym: 'gym',
  spa: 'spa',
  entertainment: 'entertainment',
  restaurant: 'restaurant',
  juicebar: 'juicebar',
  event_hall: 'event_hall',
  fitness_program: 'fitness_program',
  youth_program: 'youth_program',
} as const;

export type ListBlogPostsParams = {
category?: string;
limit?: number;
};

export type ListGalleryImagesParams = {
category?: string;
};

export type ListContactMessagesParams = {
isRead?: boolean;
};

export type GetRecentActivityParams = {
limit?: number;
};