# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ivory Health Club API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled]
        - name: serviceType
          in: query
          schema:
            type: string
            enum: [gym, spa, entertainment, restaurant, juicebar, event_hall, fitness_program, youth_program]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a service or event hall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (admin confirm/cancel)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found

  # Blog Posts
  /blog-posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List all published blog posts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
            default: 10
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a new blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Blog post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog-posts/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Blog post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostPatch"
      responses:
        "200":
          description: Blog post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # Gallery
...
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact form message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
        "400":
          description: Validation error
  /contact-messages:
    get:
      operationId: listContactMessages
      tags: [contact]
      summary: List all contact messages (admin)
      parameters:
        - name: isRead
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of contact messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
  /contact-messages/{id}/read:
    patch:
      operationId: markContactRead
      tags: [contact]
      summary: Mark a contact message as read (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadInput"
      responses:
        "200":
          description: Message updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"

  # Admin Dashboard
...
components:
  schemas:
...
    Booking:
      type: object
      required: [id, serviceType, firstName, lastName, email, phone, bookingDate, status, createdAt]
      properties:
        id:
          type: number
        serviceType:
          type: string
          enum: [gym, spa, entertainment, restaurant, juicebar, event_hall, fitness_program, youth_program]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        bookingDate:
          type: string
        bookingTime:
          type: ["string", "null"]
        numberOfGuests:
          type: ["number", "null"]
        specialRequests:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    BookingInput:
      type: object
      required: [serviceType, firstName, lastName, email, phone, bookingDate]
      properties:
        serviceType:
          type: string
          enum: [gym, spa, entertainment, restaurant, juicebar, event_hall, fitness_program, youth_program]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        bookingDate:
          type: string
        bookingTime:
          type: string
        numberOfGuests:
          type: number
        specialRequests:
          type: string
    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        notes:
          type: string
    BlogPost:
      type: object
      required: [id, title, slug, content, category, published, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        author:
          type: string
        published:
          type: boolean
        createdAt:
          type: string
    BlogPostInput:
      type: object
      required: [title, content, category, author]
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        author:
          type: string
        published:
          type: boolean
    BlogPostPatch:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        author:
          type: string
        published:
          type: boolean
...
    ContactMessage:
      type: object
      required: [id, name, email, subject, message, isRead, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
    ContactInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
    MarkReadInput:
      type: object
      required: [isRead]
      properties:
        isRead:
          type: boolean
...