# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ivory Health Club API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  author?: string;
  published: boolean;
  createdAt: string;
}