# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ivory Health Club API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingServiceType } from './bookingServiceType';
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  serviceType: BookingServiceType;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  bookingDate: string;
  /** @nullable */
  bookingTime?: string | null;
  /** @nullable */
  numberOfGuests?: number | null;
  /** @nullable */
  specialRequests?: string | null;
  status: BookingStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}