# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ivory Health Club API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingInputServiceType } from './bookingInputServiceType';

export interface BookingInput {
  serviceType: BookingInputServiceType;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  bookingDate: string;
  bookingTime?: string;
  numberOfGuests?: number;
  specialRequests?: string;
}