# lib/api-zod/src/generated/types/bookingStatusUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ivory Health Club API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatusUpdateStatus } from './bookingStatusUpdateStatus';

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
  notes?: string;
}