# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ivory Health Club API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
}